+++ src/cls/__init__.py
@@ -1,0 +1,5 @@
+from .AdsPostsWindow import AdsPostsWindow
+from .GroupPostsWindow import GroupPostsWindow
+from .MenuBar import MenuBar
+from .HomemadeApplication import HomemadeApplication
+from .ExtractDataWindow import ExtractDataWindow