+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# Ver. 0.2.2
+- Add feature extracting historical data.
+
+# Ver. 0.2.1
+- Change paradigm to OOP.
+- Add Username field.
+- Fix not filter out old users while scrape groups.
+
+# Ver. 0.2.0
+- Add feature crawl groups.
+
+# Ver. 0.1.3
+- Better catch ads posts.
+- Fix not update current old users list.
+- Catch phone numbers with 1800, 1900 head.
+
+# Ver. 0.1.2 HOTFIX
+- Fix not filter out old users.
+
 # Ver. 0.1.1 HOTFIX
 - Fix refreshing loop after log-in FB on slow computer/connection.
 - Add more ways to get page link.